# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "host": "localhost/ecotex",
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Authenticate user",
                "description": "Authenticate user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful authenticated",
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Invalid credential",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Register user",
                "description": "Register user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Users"
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Users"
                        },
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password confirmation",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Users"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful registered",
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Invalid Information",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Users": {
            "required": [
                "email",
                "type",
                "first_name",
                "last_name"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "email": {
                    "description": "email",
                    "type": "string"
                },
                "password": {
                    "description": "password",
                    "type": "string"
                },
                "type": {
                    "description": "type",
                    "type": "string"
                },
                "first_name": {
                    "description": "first_name",
                    "type": "string"
                },
                "last_name": {
                    "description": "last_name",
                    "type": "string"
                },
                "phone": {
                    "description": "phone",
                    "type": "string"
                },
                "country": {
                    "description": "country",
                    "type": "string"
                },
                "company_name": {
                    "description": "company_name",
                    "type": "string"
                },
                "brief_name": {
                    "description": "brief_name",
                    "type": "string"
                },
                "company_address": {
                    "description": "company_address",
                    "type": "string"
                },
                "website": {
                    "description": "website",
                    "type": "string"
                },
                "description": {
                    "description": "description",
                    "type": "string"
                },
                "is_paid": {
                    "description": "is_paid",
                    "type": "integer",
                    "format": "int32"
                },
                "address": {
                    "description": "address",
                    "type": "string"
                },
                "identity_card": {
                    "description": "identity_card",
                    "type": "string"
                },
                "establishment_year": {
                    "description": "establishment_year",
                    "type": "integer",
                    "format": "int32"
                },
                "business_registration_number": {
                    "description": "business_registration_number",
                    "type": "integer",
                    "format": "int32"
                },
                "form_of_ownership": {
                    "description": "form_of_ownership",
                    "type": "string"
                },
                "number_of_employees": {
                    "description": "number_of_employees",
                    "type": "integer",
                    "format": "int32"
                },
                "floor_area": {
                    "description": "floor_area",
                    "type": "number",
                    "format": "double"
                },
                "area_of_factory": {
                    "description": "area_of_factory",
                    "type": "number",
                    "format": "double"
                },
                "commercial_service_type": {
                    "description": "commercial_service_type",
                    "type": "string"
                },
                "revenue_per_year": {
                    "description": "revenue_per_year",
                    "type": "number",
                    "format": "double"
                },
                "pieces_per_year": {
                    "description": "pieces_per_year",
                    "type": "integer",
                    "format": "int32"
                },
                "compliance": {
                    "description": "compliance",
                    "type": "string"
                },
                "activation_code": {
                    "description": "activation_code",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}